Refactor into calibration.liquid module
